[Feat] Yhzhang/update mimic-it dataset (#128)
* update mimicit

* update mimic-it

* [Format Update] linted

---------

Update README.md

Fixed the display problem and table format problem of the transparent bottom of the picture in GitHub dark mode
Merge pull request #130 from Luodian/README.md-on-mimic-it

Update README.md
Update README.md of MIMIC-IT

add link to the instruction of DC and LA
Merge pull request #131 from Luodian/DC-LA-ins

Update README.md of MIMIC-IT
✨ Video interactive demo initial release

lint

Merge pull request #132 from Luodian/video_demo

✨ Video interactive demo initial release
Update README.md

Fix the link of mimic-it dataset.
Merge pull request #133 from Luodian/mimic-it-dataset-link

Update README.md
[Fix] import bugs and update wqpip packages (#135)

[Docs] Update README.md
Update README.md
Update README.md
Update README.md
Update README.md

[Doc] update readme

Merge pull request #132 from Luodian/video_demo

✨ Video interactive demo initial release

Merge branch 'main' of https://github.com/Luodian/Otter into main
