+++ pipeline/__init__.py
@@ -1,2 +1,1 @@
-from .src.collie import Flamingo
-from .src.factory import create_model_and_transforms
+
